Fluently operational

+++ update_arch-worker.sh
@@ -9,9 +9,9 @@ PACMAN_DEFAULT_CONFIG_FILE="/etc/pacman.conf"
 REPO_DIR="$(dirname "$(readlink --canonicalize "$0")")"
 BACKUP_DIR="${REPO_DIR}/backup"
 mkdir --parents "${BACKUP_DIR}"
+CUSTOM_LOG_DIR="${REPO_DIR}/logs"
 BACKUP_TIME_AND_DATE=$(cat "${CUSTOM_LOG_DIR}/update_arch-last_time_the_update_was_initiated.log")
 sudo mv "${PACMAN_DEFAULT_CONFIG_FILE}" "${BACKUP_DIR}/pacman.conf-${BACKUP_TIME_AND_DATE}.bak"
-
 
 CUSTOM_CONFIG_DIR="${REPO_DIR}/config"
 PACMAN_CUSTOM_CONFIG="${CUSTOM_CONFIG_DIR}/pacman.conf"
@@ -22,6 +22,14 @@ sudo cp --force "${PACMAN_CUSTOM_CONFIG}" "$PACMAN_DEFAULT_CONFIG_FILE"
 
 
 
+
+
+PIKAUR_CUSTOM_CONFIG="${CUSTOM_CONFIG_DIR}/pikaur.conf"
+mv "${HOME}/.config/pikaur.conf" "${BACKUP_DIR}/pikaur.conf-${BACKUP_TIME_AND_DATE}.bak"
+cp --force "$PIKAUR_CUSTOM_CONFIG" "${HOME}/.config/pikaur.conf"
+
+PACMAN_LOG_FILE="$(pacman --verbose 2>/dev/null | grep "Log File" | rev | cut --delimiter=' ' --fields=1 | rev)"
+wc -l "${PACMAN_LOG_FILE}" | cut -d' ' -f1 > "${CUSTOM_LOG_DIR}/update_arch-${BACKUP_TIME_AND_DATE}-pacman_log-starting_line_for_this_update_session.log"
 
 
 
@@ -53,9 +61,6 @@ if [[ -z $PIKAUR_INSTALLED ]]; then
   rm -rf /tmp/pikaur-git
 fi
 
-PIKAUR_CUSTOM_CONFIG="${CUSTOM_CONFIG_DIR}/pikaur.conf"
-mv "${HOME}/.config/pikaur.conf" "${BACKUP_DIR}/pikaur.conf-${BACKUP_TIME_AND_DATE}.bak"
-cp --force "$PIKAUR_CUSTOM_CONFIG" "${HOME}/.config/pikaur.conf"
 
 
 
@@ -66,18 +71,15 @@ cp --force "$PIKAUR_CUSTOM_CONFIG" "${HOME}/.config/pikaur.conf"
 
 # Download latest version of update script
 
-git_pull_status=$(git -C "$REPO_DIR" pull)
+#git_pull_status=$(git -C "${REPO_DIR}" pull)
 
-echo "$git_pull_status" | grep --invert-match "Already up to date."
+#echo "$git_pull_status" | grep --invert-match "Already up to date."
 
-test $? -eq 0
-
-is_local_repo_outdated="$?"
-if [[ is_local_repo_outdated -eq 0 ]]; then
-  printf "%s\n" "Repository updated, or some merge problem occured."
-  printf "%s\n" "Check that and then please, launch the script again."
-  exit
-fi
+#if [[ $? -eq 0 ]]; then
+#  printf "%s\n" "Repository updated, or some merge problem occured."
+#  printf "%s\n" "Check that and then please, launch the script again."
+#  exit
+#fi
 
 
 
@@ -105,12 +107,12 @@ fi
 
 # Copy embedded gpg configuration file to the system
 
-PACMAN_GPG_DIR="$(pacman --verbose 2--config "${PACMAN_CUSTOM_CONFIG}" 2>/dev/null | grep "GPG Dir" | rev | cut --delimiter=' ' --fields=1 | rev)"
-sudo mv "${PACMAN_GPG_DIR}gpg.conf" "${BACKUP_DIR}/gpg.conf-${BACKUP_TIME_AND_DATE}.bak"
+PACMAN_GPG_DIR="$(pacman --verbose --config "${PACMAN_CUSTOM_CONFIG}" 2>/dev/null | grep "GPG Dir" | rev | cut --delimiter=' ' --fields=1 | rev)"
 
-GPG_CUSTOM_CONFIG="${CUSTOM_CONFIG_DIR}/gpg.conf"
-sudo cp --force "${GPG_CUSTOM_CONFIG}" "${PACMAN_GPG_DIR}gpg.conf"
-
+if [ -f "${PACMAN_GPG_DIR}gpg.conf" ]
+then
+  sudo mv "${PACMAN_GPG_DIR}gpg.conf" "${BACKUP_DIR}/gpg.conf-${BACKUP_TIME_AND_DATE}.bak"
+fi
 
 
 
@@ -118,9 +120,17 @@ sudo cp --force "${GPG_CUSTOM_CONFIG}" "${PACMAN_GPG_DIR}gpg.conf"
 # Cleanup GPG keys
 #  see https://bbs.archlinux.org/viewtopic.php?pid=1837082#p1837082"
 
-sudo rm -R "$PACMAN_GPG_DIR"
-sudo rm -R /root/.gnupg/
-rm -rf ~/.gnupg/
+if [ -d "${PACMAN_GPG_DIR}" ]; then
+  sudo rm --recursive "${PACMAN_GPG_DIR}"
+fi
+
+if [ -d "/root/.gnupg/" ]; then
+  sudo rm --recursive /root/.gnupg/
+fi
+
+if [ -d "${HOME}/.gnupg/" ]; then
+  rm --recursive "${HOME}/.gnupg/"
+fi
 
 
 
@@ -128,6 +138,24 @@ rm -rf ~/.gnupg/
 # Initialize pacman keyring
 
 sudo pacman-key --init
+
+
+
+
+# Backup default generated gpg.conf and copy custom gpg.conf
+
+
+if [ -f "${PACMAN_GPG_DIR}gpg.conf" ]
+then
+  sudo mv "${PACMAN_GPG_DIR}gpg.conf" "${BACKUP_DIR}/gpg.conf-${BACKUP_TIME_AND_DATE}.bak"
+fi
+
+GPG_CUSTOM_CONFIG="${CUSTOM_CONFIG_DIR}/gpg.conf"
+sudo cp --force "${GPG_CUSTOM_CONFIG}" "${PACMAN_GPG_DIR}gpg.conf"
+
+
+
+
 
 # Add GPG keys for custom repositories and AUR packages
 
@@ -154,13 +182,15 @@ sudo pacman-key --lsign-key 5EE46C4C
 
 # Add GPG key for chaotic repository - Pedro Henrique Lara Campos - pedrohlc
 # Type: unofficial repo
-sudo pacman-key --keyserver hkp://pool.sks-keyservers.net --recv-keys 3056513887B78AEB
+sudo pacman-key --recv-keys 3056513887B78AEB
 sudo pacman-key --lsign-key 3056513887B78AEB
 
 # Add GPG key for Pedram Pourang - tsujan - required when building compton-conf AUR package - see https://aur.archlinux.org/packages/compton-conf/#pinned-742136
 # Type: AUR repo
+# Error code: 2 - gpg: no default secret key: No secret key Key not changed so no update needed.
 gpg --recv-keys BE793007AD22DF7E
-gpg --lsign-key BE793007AD22DF7E
+gpg2 --recv-keys BE793007AD22DF7E
+#gpg --lsign-key BE793007AD22DF7E
 
 
 # Uprade keyrings and additional mirrorlists for repositories
@@ -224,11 +254,11 @@ pikaur \
 pikaur \
     --sync \
     --refresh \
-    --needed \
     --noconfirm \
-    --config "$PACMAN_CUSTOM_CONFIG" \
-    --pikaur-config "$PIKAUR_CUSTOM_CONFIG" \
-  pikaur powerpill reflector rsync shellcheck
+    --needed \
+    --config "${PACMAN_CUSTOM_CONFIG}" \
+    --pikaur-config "${PIKAUR_CUSTOM_CONFIG}" \
+  pikaur
 
 
 
@@ -242,7 +272,7 @@ pikaur \
 
 # Clear pacman databases
 
-PACMAN_DB_PATH="$(pacman --verbose 2--config "${PACMAN_CUSTOM_CONFIG}" 2>/dev/null | grep "DB Path" | rev | cut --delimiter=' ' --fields=1 | rev)"
+PACMAN_DB_PATH="$(pacman --verbose --config "${PACMAN_CUSTOM_CONFIG}" 2>/dev/null | grep "DB Path" | rev | cut --delimiter=' ' --fields=1 | rev)"
 sudo rm -rf "$PACMAN_DB_PATH"/sync/*
 
 
@@ -252,7 +282,7 @@ sudo rm -rf "$PACMAN_DB_PATH"/sync/*
 
 # Updating official packages
 
-sudo exo-open --launch TerminalEmulator --geometry=240x24 --display :0.0 --show-menubar --show-borders --hide-toolbar --hold --command="pacman --sync --refresh --refresh --sysupgrade --needed --verbose --noconfirm" 2>&1
+sudo exo-open --launch TerminalEmulator --geometry=240x24 --display :0.0 --show-menubar --show-borders --hide-toolbar --command="pacman --sync --refresh --refresh --sysupgrade --needed --verbose --noconfirm" 2>&1 &
 
 
 
@@ -283,7 +313,7 @@ done
 
 # Updating unofficial - AUR - packages
 
-sudo exo-open --launch TerminalEmulator --geometry=240x24 --display :0.0 --show-menubar --show-borders --hide-toolbar --hold --command="pikaur --sync --refresh --refresh --sysupgrade --verbose --noedit --nodiff --noconfirm --overwrite /usr/lib/p11-kit-trust.so --overwrite /usr/bin/fwupdate --overwrite /usr/share/man/man1/fwupdate.1.gz" 2>&1
+sudo exo-open --launch TerminalEmulator --geometry=240x24 --display :0.0 --show-menubar --show-borders --hide-toolbar --command="pikaur --sync --refresh --refresh --sysupgrade --verbose --noedit --nodiff --noconfirm --overwrite /usr/lib/p11-kit-trust.so --overwrite /usr/bin/fwupdate --overwrite /usr/share/man/man1/fwupdate.1.gz" 2>&1 &
 
 
 
@@ -312,6 +342,12 @@ done
 
 
 
+# Wait for the pikaur window to close after done synchronizing databases and dropping the lock
+
+sleep 5
+
+
+
 
 
 
@@ -320,107 +356,6 @@ done
 
 rm -rf ~/.libvirt
 
-# Remove orphaned packages
-
-# Step 1: Remove implicitly installed orphaned packages
-
-# Backup pacman output for uninstallation of all orphaned packages
-
-#-s, --recursive
-#    Remove each target specified including all of their dependencies, provided that (A) they are not required by
-#    other packages; and (B) they were not explicitly installed by the user. This operation is recursive and analogous
-#    to a backwards --sync operation, and it helps keep a clean system without orphans. ...
-
-yes n | sudo pacman --remove --nosave --recursive $(sudo pacman --query --deps --unrequired --quiet) > "${CUSTOM_LOG_DIR}/update_arch-${BACKUP_TIME_AND_DATE}-orphaned_packages-implicitly_installed_as_dependencies-pacman_output.log" 2>/dev/null
-
-
-
-
-
-
-# Save all orphaned packages that were installed implicitly (automatically), as a dependency for other packages
-
-sudo pacman --query --deps --unrequired --quiet > "${CUSTOM_LOG_DIR}/update_arch-${BACKUP_TIME_AND_DATE}-orphaned_packages-implicitly_installed_as_dependencies.log"
-
-
-
-
-
-# Uninstall all orphaned packages that were installed implicitly (automatically), as a dependency for other packages
-
-number_of_implicitly_installed_dependencies="$(wc -l "${CUSTOM_LOG_DIR}/update_arch-${BACKUP_TIME_AND_DATE}-orphaned_packages-implicitly_installed_as_dependencies.log" | cut --delimiter=' ' --fields=1)"
-
-if [ $number_of_implicitly_installed_dependencies -ge 1 ]
-then
-  sudo pacman --remove --nosave --recursive --noconfirm $(cat "${CUSTOM_LOG_DIR}/update_arch-${BACKUP_TIME_AND_DATE}-orphaned_packages-implicitly_installed_as_dependencies.log")
-fi
-
-
-
-# Install back optionally required orphaned packages
-
-grep 'optionally requires' "${CUSTOM_LOG_DIR}/update_arch-${BACKUP_TIME_AND_DATE}-orphaned_packages-implicitly_installed_as_dependencies-pacman_output.log" | tr -d ':' | sed 's/^\s*//g' | cut --delimiter=' ' --fields=4 > "${CUSTOM_LOG_DIR}/update_arch-${BACKUP_TIME_AND_DATE}-orphaned_packages-implicitly_installed_as_dependencies-still_required_as_optional_dependencies_for_other_packages.log"
-
-pikaur --sync --refresh --refresh --needed --noedit --nodiff --noconfirm $(cat "${CUSTOM_LOG_DIR}/update_arch-${BACKUP_TIME_AND_DATE}-orphaned_packages-implicitly_installed_as_dependencies-still_required_as_optional_dependencies_for_other_packages.log")
-
-
-
-
-
-
-# Step 2: Remove explicitly installed orphaned packages
-
-# Save pacman output for uninstallation of all orphaned packages, even the explicitly (manually) installed
-
-#REMOVE OPTIONS (APPLY TO -R)
-#...
-#-s, --recursive
-#    ... if you want to omit condition (B), pass this option twice.
-
-yes n | sudo pacman --remove --nosave --recursive --recursive $(sudo pacman --query --deps --unrequired --quiet) > "${CUSTOM_LOG_DIR}/update_arch-${BACKUP_TIME_AND_DATE}-orphaned_packages-explicitly_installed-pacman_output.log" 2>/dev/null
-
-
-
-
-
-
-# Save all orphaned packages that were explicitly (manually) installed
-
-head -n -2 "${CUSTOM_LOG_DIR}/update_arch-${BACKUP_TIME_AND_DATE}-orphaned_packages-explicitly_installed-pacman_output.log" | tail -n +5 | tr --squeeze-repeats '[:space:]' | cut --delimiter=' ' --fields=1 > "${CUSTOM_LOG_DIR}/update_arch-${BACKUP_TIME_AND_DATE}-orphaned_packages_explicitly_installed.log"
-
-
-
-
-
-
-
-# Uninstall all orphaned packages that were installed explicitly (manually), as a dependency for other packages
-
-number_of_explicitly_installed_dependencies="$(wc -l "${CUSTOM_LOG_DIR}/update_arch-${BACKUP_TIME_AND_DATE}-orphaned_packages_explicitly_installed.log" | cut --delimiter=' ' --fields=1)"
-
-if [ $number_of_explicitly_installed_dependencies -ge 1 ]
-then
-  sudo pacman --remove --nosave --recursive --noconfirm $(cat "${CUSTOM_LOG_DIR}/update_arch-${BACKUP_TIME_AND_DATE}-orphaned_packages_explicitly_installed.log")
-fi
-
-
-
-
-
-
-
-
-
-# Install back optionally required explicitly installed orphaned packages
-
-grep 'optionally requires' "${CUSTOM_LOG_DIR}/update_arch-${BACKUP_TIME_AND_DATE}-orphaned_packages-explicitly_installed-pacman_output.log" | tr -d ':' | sed 's/^\s*//g' | cut --delimiter=' ' --fields=4 > "${CUSTOM_LOG_DIR}/update_arch-${BACKUP_TIME_AND_DATE}-orphaned_packages-explicitly_installed-still_required_as_optional_dependencies_for_other_packages.log"
-
-number_of_explicitly_installed_optional_dependencies="$(wc -l "${CUSTOM_LOG_DIR}/update_arch-${BACKUP_TIME_AND_DATE}-orphaned_packages-explicitly_installed-still_required_as_optional_dependencies_for_other_packages.log" | cut --delimiter=' ' --fields=1)"
-
-if [ $number_of_explicitly_installed_optional_dependencies -ge 1 ]
-then
-  pikaur --sync --refresh --refresh --needed --noedit --nodiff --noconfirm $(cat "${CUSTOM_LOG_DIR}/update_arch-${BACKUP_TIME_AND_DATE}-orphaned_packages-explicitly_installed-still_required_as_optional_dependencies_for_other_packages.log")
-fi
 
 
 
@@ -439,10 +374,10 @@ fi
 
 
 log_line_number_begin="$(cat "${CUSTOM_LOG_DIR}/update_arch-${BACKUP_TIME_AND_DATE}-pacman_log-starting_line_for_this_update_session.log")"
-PACMAN_LOG_FILE="$(pacman --verbose 2--config "${PACMAN_CUSTOM_CONFIG}" 2>/dev/null | grep "Log File" | rev | cut --delimiter=' ' --fields=1 | rev)"
 log_line_number_end="$(wc -l "$PACMAN_LOG_FILE" | cut -d' ' -f1)"
 update_log_lines=$(( log_line_number_end - log_line_number_begin ))
-tail -n "$update_log_lines" "${PACMAN_LOG_FILE}" | less
+
+printf "%s\n" "tac "${PACMAN_LOG_FILE}" | head --lines "${update_log_lines}" | less"
